modules/testing_qa/ci_cd.py
"""
CI/CD Integration Module

Provides CIIntegration, GitHubActions, Jenkins, and CircleCI integrations.
"""

import logging
import yaml
import json
from typing import Dict, Any, List, Optional
from pathlib import Path

logger = logging.getLogger(__name__)


class GitHubActions:
    """
    GitHub Actions CI/CD integration.

    Generates GitHub Actions workflow configurations.
    """

    def __init__(self):
        """Initialize GitHub Actions integration."""
        self.logger = logging.getLogger(__name__)

    def generate_workflow(
        self,
        name: str = "Test Suite",
        python_versions: List[str] = None,
        on_events: List[str] = None,
    ) -> Dict[str, Any]:
        """
        Generate GitHub Actions workflow configuration.

        Args:
            name: Workflow name
            python_versions: Python versions to test
            on_events: Trigger events

        Returns:
            Workflow configuration
        """
        python_versions = python_versions or ["3.9", "3.10", "3.11"]
        on_events = on_events or ["push", "pull_request"]

        workflow = {
            "name": name,
            "on": on_events,
            "jobs": {
                "test": {
                    "runs-on": "ubuntu-latest",
                    "strategy": {
                        "matrix": {
                            "python-version": python_versions,
                        }
                    },
                    "steps": [
                        {
                            "name": "Checkout code",
                            "uses": "actions/checkout@v3",
                        },
                        {
                            "name": "Set up Python",
                            "uses": "actions/setup-python@v4",
                            "with": {
                                "python-version": "${{ matrix.python-version }}",
                            }
                        },
                        {
                            "name": "Install dependencies",
                            "run": "pip install -r requirements.txt",
                        },
                        {
                            "name": "Run tests",
                            "run": "pytest --cov --junitxml=junit.xml",
                        },
                        {
                            "name": "Upload coverage",
                            "uses": "codecov/codecov-action@v3",
                        },
                        {
                            "name": "Publish test results",
                            "uses": "EnricoMi/publish-unit-test-result-action@v2",
                            "if": "always()",
                            "with": {
                                "files": "junit.xml",
                            }
                        }
                    ]
                }
            }
        }

        return workflow

    def save_workflow(
        self,
        workflow: Dict[str, Any],
        output_file: str = ".github/workflows/test.yml",
    ) -> str:
        """
        Save workflow to file.

        Args:
            workflow: Workflow configuration
            output_file: Output file path

        Returns:
            Output file path
        """
        Path(output_file).parent.mkdir(parents=True, exist_ok=True)

        with open(output_file, "w") as f:
            yaml.dump(workflow, f, default_flow_style=False, sort_keys=False)

        self.logger.info(f"GitHub Actions workflow saved: {output_file}")
        return output_file


class Jenkins:
    """
    Jenkins CI/CD integration.

    Generates Jenkins pipeline configurations.
    """

    def __init__(self):
        """Initialize Jenkins integration."""
        self.logger = logging.getLogger(__name__)

    def generate_pipeline(
        self,
        stages: List[str] = None,
    ) -> str:
        """
        Generate Jenkinsfile pipeline.

        Args:
            stages: Pipeline stages

        Returns:
            Jenkinsfile content
        """
        stages = stages or ["Build", "Test", "Deploy"]

        pipeline = """
pipeline {
    agent any

    environment {
        PYTHON_VERSION = '3.10'
    }

    stages {
"""

        for stage in stages:
            if stage == "Build":
                pipeline += """
        stage('Build') {
            steps {
                sh 'python -m venv venv'
                sh '. venv/bin/activate && pip install -r requirements.txt'
            }
        }
"""
            elif stage == "Test":
                pipeline += """
        stage('Test') {
            steps {
                sh '. venv/bin/activate && pytest --cov --junitxml=junit.xml'
            }
            post {
                always {
                    junit 'junit.xml'
                }
            }
        }
"""
            elif stage == "Deploy":
                pipeline += """
        stage('Deploy') {
            when {
                branch 'main'
            }
            steps {
                echo 'Deploying application...'
            }
        }
"""

        pipeline += """
    }

    post {
        always {
            cleanWs()
        }
    }
}
"""

        return pipeline

    def save_pipeline(
        self,
        pipeline: str,
        output_file: str = "Jenkinsfile",
    ) -> str:
        """Save Jenkinsfile."""
        with open(output_file, "w") as f:
            f.write(pipeline)

        self.logger.info(f"Jenkinsfile saved: {output_file}")
        return output_file


class CircleCI:
    """
    CircleCI integration.

    Generates CircleCI configuration.
    """

    def __init__(self):
        """Initialize CircleCI integration."""
        self.logger = logging.getLogger(__name__)

    def generate_config(
        self,
        python_version: str = "3.10",
    ) -> Dict[str, Any]:
        """
        Generate CircleCI config.

        Args:
            python_version: Python version

        Returns:
            CircleCI configuration
        """
        config = {
            "version": 2.1,
            "jobs": {
                "test": {
                    "docker": [
                        {"image": f"cimg/python:{python_version}"}
                    ],
                    "steps": [
                        "checkout",
                        {
                            "run": {
                                "name": "Install dependencies",
                                "command": "pip install -r requirements.txt",
                            }
                        },
                        {
                            "run": {
                                "name": "Run tests",
                                "command": "pytest --cov --junitxml=junit.xml",
                            }
                        },
                        {
                            "store_test_results": {
                                "path": "junit.xml"
                            }
                        }
                    ]
                }
            },
            "workflows": {
                "test-workflow": {
                    "jobs": ["test"]
                }
            }
        }

        return config

    def save_config(
        self,
        config: Dict[str, Any],
        output_file: str = ".circleci/config.yml",
    ) -> str:
        """Save CircleCI config."""
        Path(output_file).parent.mkdir(parents=True, exist_ok=True)

        with open(output_file, "w") as f:
            yaml.dump(config, f, default_flow_style=False)

        self.logger.info(f"CircleCI config saved: {output_file}")
        return output_file


class CIIntegration:
    """
    Unified CI/CD integration manager.

    Provides unified interface for multiple CI/CD platforms.
    """

    def __init__(self):
        """Initialize CI integration."""
        self.github_actions = GitHubActions()
        self.jenkins = Jenkins()
        self.circleci = CircleCI()
        self.logger = logging.getLogger(__name__)

    def setup_ci(
        self,
        platform: str,
        **kwargs,
    ) -> str:
        """
        Setup CI/CD for specified platform.

        Args:
            platform: CI/CD platform (github, jenkins, circleci)
            **kwargs: Platform-specific arguments

        Returns:
            Output file path
        """
        if platform == "github":
            workflow = self.github_actions.generate_workflow(**kwargs)
            return self.github_actions.save_workflow(workflow)

        elif platform == "jenkins":
            pipeline = self.jenkins.generate_pipeline(**kwargs)
            return self.jenkins.save_pipeline(pipeline)

        elif platform == "circleci":
            config = self.circleci.generate_config(**kwargs)
            return self.circleci.save_config(config)

        else:
            raise ValueError(f"Unknown CI platform: {platform}")

    def get_test_status(
        self,
        platform: str,
        build_id: str,
    ) -> Dict[str, Any]:
        """
        Get test status from CI platform.

        Args:
            platform: CI platform
            build_id: Build ID

        Returns:
            Build status
        """
        # Placeholder for CI API integration
        return {
            "platform": platform,
            "build_id": build_id,
            "status": "running",
            "message": "CI API integration not implemented",
        }
