modules/testing_qa/requirements.txt
# NEXUS Testing & QA Module Dependencies

# Core Testing Frameworks
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-timeout==2.2.0
pytest-xdist==3.5.0
pytest-mock==3.12.0

# Web Testing
selenium==4.15.2
playwright==1.40.0
httpx==0.25.2
requests==2.31.0

# Load Testing
locust==2.19.1

# API Testing
jsonschema==4.20.0
pydantic==2.5.2
httpx[http2]==0.25.2

# Database Testing
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Code Coverage
coverage==7.3.4
pytest-cov==4.1.0

# Security Testing
bandit==1.7.5
safety==2.3.5
pycryptodome==3.19.0

# Test Data Generation
faker==20.1.0

# Visual Testing
Pillow==10.1.0
opencv-python==4.8.1.78

# Mutation Testing
mutpy==0.7.0

# AI Integration
anthropic==0.7.7

# Web Framework
fastapi==0.105.0
uvicorn[standard]==0.25.0

# Background Tasks
celery[redis]==5.3.4
redis==5.0.1

# Dashboard
streamlit==1.29.0
plotly==5.18.0
pandas==2.1.4

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
click==8.1.7

# Reporting
allure-pytest==2.13.2
junitparser==3.1.2

# CI/CD Integration
python-gitlab==4.2.0

# Logging and Monitoring
loguru==0.7.2
sentry-sdk==1.39.1

# Type Checking
mypy==1.7.1
types-requests==2.31.0.10
